requirements.txt
torch
torchvision
transformers
modelscope
TensorFlow
keras
opencv-python
git+https://github.com/openai/CLIP.git
matplotlib
pycocotools
timm
SentencePiece
tf_slim
tf_keras
pyclipper
shapely
pillow
pyautogui
numpy
clip-by-openai